LPS-123128 Move OrderedServiceTrackerMap into another package
