+++ modules/dxp/apps/saml/saml-opensaml-integration/src/main/java/com/liferay/saml/opensaml/integration/internal/service/tracker/collections/OrderedServiceTrackerMapFactory.java
@@ -12,7 +12,7 @@
  *
  */
 
-package com.liferay.saml.opensaml.integration.internal;
+package com.liferay.saml.opensaml.integration.internal.service.tracker.collections;
 
 import java.util.function.Supplier;
